Change binary to symlink to cli.js
The core issue is that using the node shebang in bin/latexmarkdown fails
because Node.js does not accept the lack of an extension on the file, at
least not for a module file.

Using a bash shebang for bin/latexmarkdown did not work, because the
computation of the folder in which the bash script is, may not find the
right path relative to cli.js.

For instance, in installations, it found
/usr/local/node/bin/latexmarkdown, which is a symlink to
/usr/local/node/lib/node_modules/latexmarkdown/bin/latexmarkdown.
But running ../cli.js caused Node.js to try to find it at
/usr/local/node/cli.js, whilst it is located at
/usr/local/node/lib/node_modules/latexmarkdown/cli.js.

However, Node.js looks at the filename after computing redirections.

Thus, we use instead another symlink from bin/latexmarkdown to
../cli.js.