Fix binary link & upgrade dependencies
When we first implemented the binary
(installable with `npm i -g latexmarkdown`),
the Node.js ECMAScript module systems were fresh paint,
so it required a hack using a POSIX link.
That hack no longer works, but on the positive side,
the module handling is better.
We now use a .js extension in the bin/ file,
which is required for it to be recognized as a module,
and we indicate in the package.json that the binary name is the same
as the package name.

We also take the opportunity to upgrade dependencies,
which seems required to ensure the CLI tests still work.